backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
langchain
langchain-community
langchain-groq